+++ packages/platforms/browser/lib/backgrounding-listener.ts
@@ -1,6 +1,7 @@
 import {
   type BackgroundingListener,
-  type BackgroundingListenerCallback
+  type BackgroundingListenerCallback,
+  type BackgroundingListenerState
 } from '@bugsnag/core-performance'
 
 interface DocumentForVisibilityState {
@@ -8,28 +9,53 @@ interface DocumentForVisibilityState {
   visibilityState: string
 }
 
-export default function createBrowserBackgroundingListener (document: DocumentForVisibilityState) {
+interface WindowWithDocumentForVisibilityState {
+  document: DocumentForVisibilityState
+  addEventListener: (event: string, callback: () => void) => void
+}
+
+export default function createBrowserBackgroundingListener (window: WindowWithDocumentForVisibilityState) {
   const callbacks: BackgroundingListenerCallback[] = []
+  let state: BackgroundingListenerState = window.document.visibilityState === 'hidden'
+    ? 'in-background'
+    : 'in-foreground'
 
   const backgroundingListener: BackgroundingListener = {
     onStateChange (backgroundingListenerCallback: BackgroundingListenerCallback): void {
       callbacks.push(backgroundingListenerCallback)
 
       // trigger the callback immediately if the document is already 'hidden'
-      if (document.visibilityState === 'hidden') {
-        backgroundingListenerCallback('in-background')
+      if (state === 'in-background') {
+        backgroundingListenerCallback(state)
       }
     }
   }
 
-  document.addEventListener('visibilitychange', function () {
-    const state = document.visibilityState === 'hidden'
-      ? 'in-background'
-      : 'in-foreground'
+  const backgroundStateChanged = (newState: BackgroundingListenerState) => {
+    if (state === newState) return
 
+    state = newState
     for (const callback of callbacks) {
       callback(state)
     }
+  }
+
+  window.document.addEventListener('visibilitychange', function () {
+    const newState = window.document.visibilityState === 'hidden'
+      ? 'in-background'
+      : 'in-foreground'
+
+    backgroundStateChanged(newState)
+  })
+
+  // some browsers don't fire the visibilitychange event when the page is suspended,
+  // so we also listen for pagehide and pageshow events
+  window.addEventListener('pagehide', function () {
+    backgroundStateChanged('in-background')
+  })
+
+  window.addEventListener('pageshow', function () {
+    backgroundStateChanged('in-foreground')
   })
 
   return backgroundingListener