+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - (browser) Fall back to unbuffered performance observer when buffer is not supported [#352](https://github.com/bugsnag/bugsnag-js-performance/pull/352)
+- (browser) Listen for pagehide and pageshow events in backgrounding listener [#362](https://github.com/bugsnag/bugsnag-js-performance/pull/362)
 
 ## v1.2.0 (2023-10-12)